pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "wok"
version = "0.1.0"
description = "Wok is a tool to manage several `git` repos in a single workspace."
authors = ["Serge Matveenko <lig@countzero.co>"]
license = "MIT"

[tool.poetry.scripts]
wok = 'wok.cli:main'

[tool.poetry.dependencies]
# For pygit2 dependencies see https://github.com/python-poetry/poetry/issues/2066
attrs = "^19.3.0"
cached-property = "^1.5.1" # pygit2 dependency
cffi = "^1.14.0" # pygit2 dependency
click = "^7.0"
marshmallow = "^3.5.1"
pygit2 = "^1.1.1"
python = "^3.7"
pyyaml = "^5.3.1"
toml = "^0.10.0"

[tool.poetry.dev-dependencies]
pytest = "^5.3.5"
pytest-black = "^0.3.7"
pytest-flake8 = "^1.0.4"
pytest-isort = "^0.3.1"
pytest-mypy = "^0.4.2"
sphinx = "^2.4.4"

[tool.black]
target-version = ["py37"]
skip-string-normalization = true

[tool.isort]
force_grid_wrap = 0
include_trailing_comma = true
line_length = 88
multi_line_output = 3
use_parentheses = true
lines_after_imports = 2
